api/functions/generate/requirements.txt
tiktoken
boto3
faiss-cpu
requests
numpy
nltk 
pdfplumber